MCR-3817 Keep the webcli playwright setup in its own module
Playwright transpiles a .ts config to CommonJS and applies the paths of the
app tsconfig.json, so @playwright/test resolves against the shared node_modules
of mycore-vue. Only the .mts variant goes through the real ES module loader,
which ignores those paths. Move both configs back next to the tests they run
and keep the extension at .ts.

The websocket stub of the performance probe is started by playwright as a plain
node process, so it gets neither the transpilation nor the paths. It now pulls
ws through a createRequire anchored in mycore-vue, the same idea as the anchor
in vite.shared.ts, and moves back to the app as well.
