+++ mycore-vue/README.md
@@ -25,9 +25,6 @@ and the viewer) are not part of this module. They keep their own `package.json`,
 | `tsconfig.json` | shared TypeScript compiler options, extended by the app configs |
 | `env.d.ts` | shared ambient types, referenced by the `env.d.ts` of every app |
 | `eslint.config.mjs` | shared flat eslint config, re-exported by the `eslint.config.mjs` of an app |
-| `playwright.webcli.config.mts` | playwright setup of the webcli accessibility tests |
-| `playwright.webcli.performance.config.mts` | playwright setup of the webcli performance probe |
-| `testing/` | test servers that need a toolchain dependency, e.g. the webcli websocket stub |
 
 ## Adding an app to the shared toolchain
 
@@ -139,11 +136,13 @@ only walks up the directory tree of the importing file. Both toolchains have to 
   `test:webcli`, `test-coverage:webcli`, `lint:webcli`, `typecheck:webcli`, `ci-check:webcli`,
   `test-a11y:webcli` and `test-performance:webcli`. The vitest configuration stays in the app's `vite.config.mts`, its paths are relative to the
   app root, which `defineMCRVueApp` sets.
-* The playwright configuration lives here rather than next to the tests it runs. Playwright loads a `.mts` config
-  as a real ES module, and node would not resolve `@playwright/test` from the app directory. The tests themselves
-  stay in the app. A test server that playwright starts as a plain node process follows the same rule: the webcli
-  websocket stub imports `ws` and therefore lives in `testing/`, while the a11y static server only uses node
-  builtins and stays in the app.
+* The playwright configuration stays with the app it tests, and it has to keep the extension `.ts`. Playwright
+  transpiles a `.ts` config to CommonJS and applies the `paths` of the app's `tsconfig.json`, which is what makes
+  `@playwright/test` resolve against the shared node_modules. A `.mts` config is loaded as a real ES module, and
+  that loader ignores those `paths` and fails with `ERR_MODULE_NOT_FOUND`. The same applies to the test files.
+* A test server that playwright starts as a plain node process gets neither the transpilation nor the `paths`.
+  The webcli websocket stub therefore pulls `ws` through a `createRequire` anchored in this module, the same
+  idea as the anchor in `vite.shared.ts`. The a11y static server only uses node builtins and needs nothing.
 * `vue-i18n` is pinned to the exact version `11.4.2`. From `11.4.3` on it requires node 22, while the reactor POM
   pins `node.version` to `v20.19.0`. The pin can be dropped as soon as the node version is raised.
 * `vite-plugin-eslint` (access-key-manager2) is unmaintained and is not part of the shared toolchain. Linting runs
