+++ mycore-vue/package.json
@@ -20,11 +20,11 @@
     "dev:jobqueue": "vite --config ../mycore-jobqueue/src/main/vue/jobqueue/vite.config.mts",
     "build:webcli": "vite build --config ../mycore-webcli/src/main/vue/webcli/vite.config.mts",
     "typecheck:webcli": "vue-tsc --noEmit --pretty false -p ../mycore-webcli/src/main/vue/webcli/tsconfig.typecheck.json",
-    "lint:webcli": "cd ../mycore-webcli/src/main/vue/webcli && eslint src tests vite.config.mts",
+    "lint:webcli": "cd ../mycore-webcli/src/main/vue/webcli && eslint src tests vite.config.mts playwright.config.ts playwright.performance.config.ts",
     "test:webcli": "vitest --run --config ../mycore-webcli/src/main/vue/webcli/vite.config.mts",
     "test-coverage:webcli": "vitest --run --coverage --config ../mycore-webcli/src/main/vue/webcli/vite.config.mts",
-    "test-a11y:webcli": "playwright test --config playwright.webcli.config.mts",
-    "test-performance-only:webcli": "playwright test --config playwright.webcli.performance.config.mts",
+    "test-a11y:webcli": "playwright test --config ../mycore-webcli/src/main/vue/webcli/playwright.config.ts",
+    "test-performance-only:webcli": "playwright test --config ../mycore-webcli/src/main/vue/webcli/playwright.performance.config.ts",
     "test-performance:webcli": "run-s build:webcli test-performance-only:webcli",
     "ci-check:webcli": "run-s lint:webcli typecheck:webcli test-coverage:webcli",
     "dev:webcli": "vite --config ../mycore-webcli/src/main/vue/webcli/vite.config.mts",
